+++ CHANGELOG.md
@@ -52,6 +52,8 @@ Summary:
 -   Changed `--force` to be the same as `--defaults --overwrite`.
 -   Copied files will reflect permissions on the same files in the template.
 -   Copier now uses `git clone --filter=blob:none` when cloning, to be faster.
+-   Removing files from templates will remove them too from the subprojects when they
+    get updated.
 
 ### Deprecated